Use react context
