+++ src/pages/Sample/sample.scss
@@ -11,3 +11,8 @@
     }
   }
 }
+
+.job-group {
+  margin-bottom: 40px;
+  margin-top: 20px;
+}
